requirements.txt
chess
websocket-client